fix(2_0_1_7): widen character AccountId in ACWorldListPacket to 8 bytes
Same pattern as 1904b2bc on ACAuthResponsePacket — LoginCharacterInfo
.AccountId is uint (4 bytes) but the 2.0.1.7 client decodes the
character record as ulong AccountId (8 bytes) followed by the rest of
the fields, so the layout shifts by 4 bytes per character entry.

Only triggers when the account has at least one character; with zero
characters the for-loop never runs and the packet is identical to 1.2.

+++ AAEmu.Login/Core/Packets/L2C/ACWorldListPacket.cs
@@ -56,7 +56,11 @@ public class ACWorldListPacket(List<GameServer> gameServers, List<LoginCharacter
         {
             foreach (var character in characters)
             {
-                stream.Write(character.AccountId);
+                // LoginCharacterInfo.AccountId is uint (4 bytes) — 2.0.1.7 expects
+                // ulong (8 bytes), same pattern as ACAuthResponsePacket. Cast on the
+                // wire so every following character field stays at the offset the
+                // 2.0 client expects.
+                stream.Write((ulong)character.AccountId);
                 stream.Write(character.GsId);
                 stream.Write(character.Id);
                 stream.Write(character.Name);
